todo_backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask[async]>=3.1.2",
    "gunicorn>=23.0.0",
    "nats-py>=2.12.0",
    "psycopg2-binary>=2.9.11",
]
